Fix css can not be loaded before HEAD error
The page was trying to load the css after the HEAD was loaded. The problem was the tribute.css file.
I instead used the moodle way to add css to plugins, by using the styles.css file and then the css is automatically laoded onto the page. So no need to add it in the lib.php file.