Fixes to @ tagging
The tagging didn't work in advanced editor. I added the functionality for this.

The DOMSubtreeModified event was firing a lot so added a throttle to make it not fire as much if there are a lot of DOM changes in a tiny space of time (50ms). it will just fire a max of once every 50ms rather than every 1ms which sometimes happened.

The positioning of the menu with all the names was off and getting more complicated for every text area the tagging was added to. If there was 2 textareas that it was added to on the same page it would position only of the first text editor regardless of which area text area the user was using. It also placed the name menu at the top of the textarea even if the person was using the @ tag multiple lines down the page. I changed it to be a lot of generic so we don't have to specify for each editor. It now places the menu where the cursor is so it solves all the positioning problems.